Moved Block draw inside Block
